Lab7's Exp10 update & misc Lab7 upload
Updating Exp10 functionality, namely:
A) Changed clock from C8 to B8 for consistency
B) Swapped the output of "Memory" segment of Q5 and Q6 due to a poorly generated module
C) Grafted an extra "MUXLIFE" in order to displaywhat's in the memory address, so the seven-segment can just display off ReadData.

And as mentioned, files for experiments 3, 5, and 8 were uploaded as well.